dom/nodes/aria-element-reflection.tentative.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8" />
    <title>Element Reflection for aria-activedescendant and aria-errormessage</title>
    <link rel=help href="https://whatpr.org/html/3917/common-dom-interfaces.html#reflecting-content-attributes-in-idl-attributes:element">
    <link rel="author" title="Meredith Lane" href="meredithl@chromium.org">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <div id="activedescendant" aria-activedescendant="x"></div>

  <div id="parent-listbox" role="listbox" aria-activedescendant="i1">
    <div role="option" id="i1">Item 1</div>
    <div role="option" id="i2">Item 2</div>
  </div>

  <script>
  test(function(t) {
    var ancestor = document.getElementById("parent-listbox");
    assert_equals(activedescendant.ariaActiveDescendantElement, null,
                  "invalid ID for relationship returns null");

    var descendant1 = document.getElementById("i1");
    var descendant2 = document.getElementById("i2");

    // Element reference should be set if the content attribute was included.
    assert_equals(ancestor.getAttribute("aria-activedescendant"), "i1", "check content attribute after parsing.");
    assert_equals(ancestor.ariaActiveDescendantElement, i1, "check idl attribute after parsing.");

    // If we set the content attribute, the element reference should reflect this.
    ancestor.setAttribute("aria-activedescendant", "i2");
    assert_equals(ancestor.ariaActiveDescendantElement, descendant2, "setting the content attribute updates the element reference.");

    // Setting the element reference should be reflected in the content attribute.
    ancestor.ariaActiveDescendantElement = descendant1;
    assert_equals(ancestor.ariaActiveDescendantElement, descendant1, "getter should return the right element reference.");
    assert_equals(ancestor.getAttribute("aria-activedescendant"), "i1", "content attribute should reflect the element reference.");

    // Both content and IDL attribute should be nullable.
    ancestor.ariaActiveDescendantElement = null;
    assert_equals(ancestor.ariaActiveDescendantElement, null);
    assert_false(ancestor.hasAttribute("aria-activedescendant"));
    assert_equals(ancestor.getAttribute("aria-activedescendant"), null, "nullifying the idl attribute removes the content attribute.");

    // Setting content attribute to non-existent or non compatible element should nullify the IDL attribute.
    // Reset the element to an existant one.
    ancestor.setAttribute("aria-activedescendant", "i1");
    assert_equals(ancestor.ariaActiveDescendantElement, i1, "reset attribute.");

    ancestor.setAttribute("aria-activedescendant", "non-existent-element");
    assert_equals(ancestor.getAttribute("aria-activedescendant"), "non-existent-element");
    assert_equals(ancestor.ariaActiveDescendantElement, null,"non-DOM content attribute should null the element reference");
  }, "aria-activedescendant element reflection");
  </script>

  <div id="parent-listbox-2" role="listbox" aria-activedescendant="option1">
    <div role="option" id="option1">Item 1</div>
    <div role="option" id="option2">Item 2</div>
  </div>

  <script>
  test(function(t) {
    var ancestor = document.getElementById("parent-listbox-2");
    var option1 = document.getElementById("option1");
    var option2 = document.getElementById("option2");
    assert_equals(ancestor.ariaActiveDescendantElement, option1);

    option1.removeAttribute("id");
    option2.setAttribute("id", "option1");
    var option2Duplicate = document.getElementById("option1");
    assert_equals(option2, option2Duplicate);

    assert_equals(ancestor.ariaActiveDescendantElement, option2);
  }, "Reassigning an ID reflects the correct element in the DOM.");
  </script>

  <div id="blank-id-parent" role="listbox">
    <div role="option" id="multiple-id"></div>
    <div role="option" id="multiple-id"></div>
  </div>

  <script>
  test(function(t) {
    var ancestor = document.getElementById("blank-id-parent");

    // Get second child of parent. This violates the setting of a reflected element
    // as it will not be the first child of the parent with that ID, which should
    // result in an empty string for the content attribute.
    ancestor.ariaActiveDescendantElement = ancestor.children[1];
    assert_true(ancestor.hasAttribute("aria-activedescendant"));
    assert_equals(ancestor.getAttribute("aria-activedescendant"), "");
    assert_equals(ancestor.ariaActiveDescendantElement, ancestor.children[1]);
  }, "element reflection with multiple ids");
  </script>

  <div id="outer-container">
    <p id="light-paragraph">Hello world!</p>
    <span class="shadow-host">
    </span>
  </div>

  <script>
  test(function(t) {
    const shadowElement = document.querySelector(".shadow-host");
    const shadow = shadowElement.attachShadow({mode: "open"});
    const link = document.createElement("a");
    shadow.appendChild(link);

    var lightPara = document.getElementById("light-paragraph");
    assert_equals(lightPara.ariaActiveDescendantElement, null);

    // The given element crosses a shadow dom boundary, so it cannot be
    // set as an element reference.
    lightPara.ariaActiveDescendantElement = link;
    assert_equals(lightPara.ariaActiveDescendantElement, null);

    // The given element crosses a shadow dom boundary (upwards), so
    // can be used as an element reference, but the content attribute
    // should reflect the empty string.
    link.ariaActiveDescendantElement = lightPara;
    assert_equals(link.ariaActiveDescendantElement, lightPara);
    assert_equals(link.getAttribute("aria-activedescendant"), "");
  }, "Setting an element reference that crosses into a shadow tree is disallowed, but setting one that is in a shadow inclusive ancestor is allowed.");
  </script>

  <input id="startTime" ></input>
  <span id="errorMessage">Invalid Time</span>

  <script>
  test(function(t) {
    var inputElement = document.getElementById("startTime");
    var errorMessage = document.getElementById("errorMessage");

    inputElement.ariaErrorMessageElement = errorMessage;
    assert_equals(inputElement.getAttribute("aria-errormessage"), "errorMessage");

    inputElement.ariaErrorMessageElement = null;
    assert_equals(inputElement.ariaErrorMessageElement, null, "blah");
    assert_false(inputElement.hasAttribute("aria-errormessage"));

    inputElement.setAttribute("aria-errormessage", "errorMessage");
    assert_equals(inputElement.ariaErrorMessageElement, errorMessage);

  }, "aria-errormessage");

  </script>

  <label>
    Password:
    <input id="password-field" type="password" aria-details="pw">
  </label>

  <ul>
    <li id="list-item-1">First description.</li>
    <li id="list-item-2">Second description.</li>
  </ul>

  <script>

  test(function(t) {
    var inputElement = document.getElementById("password-field");
    var ul1 = document.getElementById("list-item-1");
    var ul2 = document.getElementById("list-item-2");

    assert_equals(inputElement.ariaDetailsElement, null);
    inputElement.ariaDetailsElement = ul1;
    assert_equals(inputElement.getAttribute("aria-details"), "list-item-1");

    inputElement.ariaDetailsElement = ul2;
    assert_equals(inputElement.getAttribute("aria-details"), "list-item-2");
  }, "aria-details");
  </script>

</html>
